docs(rusmpp-core/decode): added some docs to decode error
